fixed parsing number from jsonVariable

+++ client/services/code_preprocessors/PythonCodePreprocessorService.js
@@ -50,7 +50,7 @@ tie.factory('PythonCodePreprocessorService', [
       } else if (typeof jsonVariable == 'boolean') {
         return jsonVariable ? 'True' : 'False';
       } else if (typeof jsonVariable === 'number') {
-	      return parseInt(jsonVariable);
+	      return parseFloat(jsonVariable);
       } else if (Array.isArray(jsonVariable)) {
         // We have to recursively convert the array's elements to Python variables.
         var pythonElements = jsonVariable.map(function(arrayElement) {
